Moved the module loading logic to a separate class
