+++ Source/System.Management/System.Management.csproj
@@ -500,6 +500,7 @@
     <Compile Include="Microsoft.PowerShell\Commands\RemoveModuleCommand.cs" />
     <Compile Include="Pash\Implementation\ModuleIntrinsics.cs" />
     <Compile Include="Microsoft.PowerShell\Commands\ExportModuleMemberCommand.cs" />
+    <Compile Include="Pash\Implementation\ModuleLoader.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
